feat: poc llm class router
